src/stage_updater/requirements.txt
Flask==2.1.0
google-cloud-firestore==1.9.0
google-cloud-pubsub==2.9.0
gunicorn==20.1.0
requests